# storage/engine: reenable the Seek optimization