Primarily affects conditional-put heavy transactions where there is an
existing value.

Fixes #6980.

name                    old time/op  new time/op  delta
Update1_Cockroach-8      356µs ± 2%   348µs ± 1%  -2.42%  (p=0.000 n=10+10)
Update10_Cockroach-8     647µs ± 2%   619µs ± 2%  -4.33%   (p=0.000 n=9+10)
Update100_Cockroach-8   2.92ms ± 2%  2.67ms ± 3%  -8.54%  (p=0.000 n=10+10)
Update1000_Cockroach-8  24.5ms ± 1%  22.3ms ± 2%  -8.91%  (p=0.000 n=10+10)

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="35" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/7016)
<!-- Reviewable:end -->